llm-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.29.7
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
